// gui/src/app/pages/DandisetPage/DandisetViewFromDendro/types.ts
export type DandisetSearchResultItem = {
    identifier: string
    created: string
    modified: string
    contact_person: string
    embargo_status: string
    most_recent_published_version?: {
        version: string
        name: string
        asset_count: number
        size: number
        status: string
        created: string
        modified: string
    }
    draft_version?: {
        version: string
        name: string
        asset_count: number
        size: number
        status: string
        created: string
        modified: string
    }
}